Split validation to separate units
